backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
playwright==1.47.2

